Remove sorting from sql.rs tests
Remove sorting in all UTs but one (the first one) - kept as a reference.

We use in-order B-tree traversal, so we expect a sorted (ascending) order, and we want to test that, too. Tests with large DB, "superheroes", visit all nodes (pages), using the `WHERE` clause, and we want to make sure that they are indeed visited in order (in-order).
Although, we have had the UT "select_id_from_superheroes" that serves the same purpose, but we want to make all our tests more robust, regardless.